Move options to Options folder.
