+++ src/Puzzle/DependencyInjection.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Puzzle.Abstractions;
 using Puzzle.Bootstrap;
+using Puzzle.Options;
 
 namespace Puzzle;
