+++ src/Puzzle/Plugin.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using System.Reflection;
 using Microsoft.Extensions.Configuration;
 using Puzzle.Abstractions;
+using Puzzle.Options;
 
 namespace Puzzle;
