+++ src/Puzzle/Puzzle.csproj
@@ -65,4 +65,8 @@
         </Content>
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Options\" />
+    </ItemGroup>
+
 </Project>
